more ui fixes
