--- electron/src/renderer/components/WelcomePage.jsx
@@ -1,174 +1,0 @@
-import React from "react";
-// import * as Realm from "realm-web";
-import {
-  Container,
-  TextField,
-  Button,
-  IconButton,
-  Card,
-  Typography,
-  InputAdornment,
-} from "@material-ui/core";
-import Visibility from "@material-ui/icons/Visibility";
-import VisibilityOff from "@material-ui/icons/VisibilityOff";
-import { Link, useHistory } from 'react-router-dom';
-// import { useRealmApp } from "./RealmApp";
-// import { MoreInfoTemplateAndDocs } from "./MoreInfo";
-// import { toggleBoolean } from "../utils";
-// import { useErrorAlert } from "../hooks/useErrorAlert";
-import useRealmApp from '../hooks/useRealmApp';
-
-export function WelcomePage() {
-    const { logIn, signUp } = useRealmApp();
-    const history = useHistory();
-  // Track whether the user is logging in or signing up for a new account
-  const [isSignup, setIsSignup] = React.useState(false);
-  const toggleIsSignup = () => {
-    clearErrors();
-    // setIsSignup(toggleBoolean);
-  };
-  // Authentication errors
-  const noErrors = {
-    email: null,
-    password: null,
-    other: null,
-  };
-  const [error, setError] = React.useState(noErrors);
-  const clearErrors = () => setError(noErrors);
-//   const NonAuthErrorAlert = useErrorAlert({
-//     error: error.other,
-//     clearError: () => {
-//       setError((prevError) => ({ ...prevError, other: null }));
-//     },
-//   });
-  // Manage password visibility
-  const [showPassword, setShowPassword] = React.useState(false);
-//   const toggleShowPassword = () => setShowPassword(toggleBoolean);
-
-  const onFormSubmit = async ({ email, password }) => {
-    clearErrors();
-    try {
-      if (isSignup) {
-        // await realmApp.emailPasswordAuth.registerUser(email, password);
-      }
-    //   await realmApp.logIn(Realm.Credentials.emailPassword(email, password));
-    } catch (err) {
-      handleAuthenticationError(err, setError);
-    }
-  };
-
-  return (
-    <Container maxWidth="sm" className="main-container">
-      <Card className="auth-card" variant="outlined">
-        <form
-          className="auth-form"
-          onSubmit={(e) => {
-            e.preventDefault();
-            const formData = new FormData(e.target);
-            const { email, password } = Object.fromEntries(formData.entries());
-            onFormSubmit({ email, password });
-          }}
-        >
-          <Typography component="h2" variant="h4" gutterBottom>
-            {isSignup ? "Sign Up" : "Log In"}
-          </Typography>
-          {/* <NonAuthErrorAlert /> */}
-          <TextField
-            id="input-email"
-            name="email"
-            label="Email Address"
-            variant="outlined"
-            error={Boolean(error.email)}
-            helperText={error.email ?? ""}
-          />
-          <br/>
-          <TextField
-            id="input-password"
-            type={showPassword ? "text" : "password"}
-            name="password"
-            label="Password"
-            variant="outlined"
-            error={Boolean(error.password)}
-            helperText={error.password ?? ""}
-            InputProps={{
-              endAdornment: (
-                <InputAdornment position="end">
-                  <IconButton
-                    aria-label="toggle password visibility"
-                    // onClick={toggleShowPassword}
-                    onMouseDown={(e) => {
-                      e.preventDefault();
-                    }}
-                  >
-                    {showPassword ? <Visibility /> : <VisibilityOff />}
-                  </IconButton>
-                </InputAdornment>
-              ),
-            }}
-          />
-          <button
-            type="button"
-            className="link-button"
-            onClick={() => toggleIsSignup()}
-          >
-            {isSignup
-              ? "Already have an account? Log In"
-              : "Don't have an account? Create an Account"}
-          </button>
-        </form>
-      </Card>
-    </Container>
-  );
-}
-
-function handleAuthenticationError(err, setError) {
-  const handleUnknownError = () => {
-    setError((prevError) => ({
-      ...prevError,
-      other: "Something went wrong. Try again in a little bit.",
-    }));
-    console.warn(
-      "Something went wrong with a Realm login or signup request. See the following error for details."
-    );
-    console.error(err);
-  };
-  if (err instanceof Realm.MongoDBRealmError) {
-    const { error, statusCode } = err;
-    const errorType = error || statusCode;
-    switch (errorType) {
-      case "invalid username":
-        setError((prevError) => ({
-          ...prevError,
-          email: "Invalid email address.",
-        }));
-        break;
-      case "invalid username/password":
-      case "invalid password":
-      case 401:
-        setError((prevError) => ({
-          ...prevError,
-          password: "Incorrect password.",
-        }));
-        break;
-      case "name already in use":
-      case 409:
-        setError((prevError) => ({
-          ...prevError,
-          email: "Email is already registered.",
-        }));
-        break;
-      case "password must be between 6 and 128 characters":
-      case 400:
-        setError((prevError) => ({
-          ...prevError,
-          password: "Password must be between 6 and 128 characters.",
-        }));
-        break;
-      default:
-        handleUnknownError();
-        break;
-    }
-  } else {
-    handleUnknownError();
-  }
-}
